### 2017-06-04 #277 @labarba
As explained in the README, this code enables PETSc-based application software to use the NVIDIA AmgX library, bestowing multi-GPU powers. Needless to say, testing the code will require both PETSc and AmgX, both big libraries.

Right now, you need to be a registered CUDA developer to download AmgX for free—but NVIDIA folks told me they have decided to open-source it. Great news! The bad news is that it's not been released yet. For now, you need a trial license.
