### 2017-06-04 #277 @labarba
On the list of registered reviewers, I found only one who lists GPU computing among their interests: @arghdos.
