### 2017-06-04 #277 @labarba
I wonder if @jedbrown might give reviewing for JOSS a go.
