### 2017-06-05 #277 @labarba
@jedbrown : (In case it hasn't show up in your feeds yet…) JOSS reviews involve downloading and installing the software, and inspecting the repository for key elements.
http://joss.theoj.org/about#reviewer_guidelines

Editors and reviewers post comments on the GitHub issue tracker, and authors are expected to respond to the comments and improve their submission until acceptance (or withdraw, if they feel unable to satisfy the review).
