### 2017-06-05 #280 @whedon
[REVIEW]: AmgXWrapper: An interface between PETSc and the NVIDIA AmgX library
**Submitting author:** @piyueh (<a href="http://orcid.org/0000-0001-6330-2709">Pi-Yueh Chuang</a>)
**Repository:** <a href="https://github.com/barbagroup/AmgXWrapper" target ="_blank">https://github.com/barbagroup/AmgXWrapper</a>
**Version:** v1.0
