### 2017-06-05 #280 @danielskatz
@jedbrown - this is ready for you to review.  You should review the reviewer instructions on http://joss.theoj.org/about. Let me know if you have any problems.
