### 2017-06-05 #280 @labarba
@jedbrown : I have enquired with NVIDIA about the trial license of AmgX. They are sure going to open source it, but it's undergoing code review right now.
