### 2017-06-11 aff91d0
remove extra horizontal lines
