### 2017-06-11 #280 @danielskatz
@jedbrown - I talked to @labarba last week, and I believe she's working on getting NVIDIA to provide a license for AmgX for you.
