### 2017-06-12 #280 @jedbrown
Joe Eaton sent me a license a few days ago.  Thanks.

"Daniel S. Katz" <notifications@github.com> writes:

> @jedbrown - I talked to @labarba last week, and I believe she's working on getting NVIDIA to provide a license for AmgX for you.
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/openjournals/joss-reviews/issues/280#issuecomment-307663922
