### 2017-07-17 #280 @jedbrown
Sorry, I've been delayed by our newborn, but will be able to finish this soon.

"Daniel S. Katz" <notifications@github.com> writes:

> We seem to have lost @jedbrown .  @labarba, maybe we need to find another reviewer.  Do you have more suggestions?
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/openjournals/joss-reviews/issues/280#issuecomment-315728155
