### 2017-08-04 #280 @jedbrown
I'm installing now.  Sorry about the delays.

"Daniel S. Katz" <notifications@github.com> writes:

> @jedbrown - any progress?  Or should we say this isn't going to work and find another reviewer?  (If so, any suggestions you have would be very welcome.)
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/openjournals/joss-reviews/issues/280#issuecomment-319682210
