### 2017-08-07 aa3a50b
type in install.md
