+++ doc/install.md
@@ -33,7 +33,7 @@ Run CMake.
 
 ```bash
 $ cmake \
-    -D CMAKE_INSTALL_PREFIC=${PATH_TO_WHERE_YOU_WANT_TO_INSTALL_AmgXWrapper} \
+    -D CMAKE_INSTALL_PREFIX=${PATH_TO_WHERE_YOU_WANT_TO_INSTALL_AmgXWrapper} \
     -D PETSC_DIR=${PATH_TO_PETSC} \
     -D PETSC_ARCH=${THE_BUILD_OF_PETSC_YOU_WANT_TO_USE} \
     -D CUDA_DIR=${PATH_TO_CUDA} \
