### 2017-08-07 c1c96fc
use CC and CXX variables as CMake community suggests
