### 2017-08-07 574d619
resolve FindPETSc.cmake didn't take environment variables
