### 2017-08-08 10b30ad
resolve issue #12: AmgX_CPU failure
Given that AmgX hasn't improved its CPU support, we decided to remove CPU
support of AmgX in the AmgXWrapper and all examples (i.e., the `h**I`
mode in the wrapper and `AmgX_CPU` mode in the exmaples).
