### 2017-08-08 c09412a
Examples: distinguish vector 2-norm from approx. continuous L2 norm
The Poisson discretization converges at second order in the continuous
L2 norm, but the convergence rate in the vector 2-norm depends on the
dimension of the problem.  This normalization (by vector length) yields
a norm that scales like the continuous L2 norm and thus converges at the
expected rate independent of dimension.
