+++ example/solveFromFiles/src/solve.cpp
@@ -35,7 +35,7 @@ PetscErrorCode solve(KSP &ksp, Mat &A, Vec &lhs, Vec &rhs, Vec &exact, Vec &err,
 
         PetscInt                Niters; // iterations used to converge
 
-        PetscScalar             norm2,  // L2 norm of solution errors
+        PetscScalar             norm2,  // 2 norm of solution errors
                                 normM;  // infinity norm of solution errors
 
         ierr = KSPGetConvergedReason(ksp, &reason); CHKERRQ(ierr);
@@ -53,7 +53,7 @@ PetscErrorCode solve(KSP &ksp, Mat &A, Vec &lhs, Vec &rhs, Vec &exact, Vec &err,
 
         // print infromation
         ierr = PetscPrintf(PETSC_COMM_WORLD, "\tSolve Time: %f\n", time); CHKERRQ(ierr);
-        ierr = PetscPrintf(PETSC_COMM_WORLD, "\tL2-Norm: %g\n", (double)norm2); CHKERRQ(ierr);
+        ierr = PetscPrintf(PETSC_COMM_WORLD, "\t2-Norm: %g\n", (double)norm2); CHKERRQ(ierr);
         ierr = PetscPrintf(PETSC_COMM_WORLD, "\tMax-Norm: %g\n", (double)normM); CHKERRQ(ierr);
         ierr = PetscPrintf(PETSC_COMM_WORLD, "\tIterations %D\n", Niters); CHKERRQ(ierr); 
 
@@ -135,7 +135,7 @@ PetscErrorCode solve(AmgXSolver &amgx, Mat &A, Vec &lhs, Vec &rhs, Vec &exact, V
 
         PetscInt                Niters; // iterations used to converge
 
-        PetscScalar             norm2,  // L2 norm of solution errors
+        PetscScalar             norm2,  // 2 norm of solution errors
                                 normM;  // infinity norm of solution errors
 
         ierr = amgx.getIters(Niters); CHKERRQ(ierr);
@@ -148,7 +148,7 @@ PetscErrorCode solve(AmgXSolver &amgx, Mat &A, Vec &lhs, Vec &rhs, Vec &exact, V
 
         // print infromation
         ierr = PetscPrintf(PETSC_COMM_WORLD, "\tSolve Time: %f\n", time); CHKERRQ(ierr);
-        ierr = PetscPrintf(PETSC_COMM_WORLD, "\tL2-Norm: %g\n", (double)norm2); CHKERRQ(ierr);
+        ierr = PetscPrintf(PETSC_COMM_WORLD, "\t2-Norm: %g\n", (double)norm2); CHKERRQ(ierr);
         ierr = PetscPrintf(PETSC_COMM_WORLD, "\tMax-Norm: %g\n", (double)normM); CHKERRQ(ierr);
         ierr = PetscPrintf(PETSC_COMM_WORLD, "\tIterations %D\n", Niters); CHKERRQ(ierr); 
 
