### 2017-08-08 #280 @jedbrown
I've filed issues in AmgXWrapper for the technical concerns that came up during my review and they are either resolved, in the process of being resolved, or not blocking issues.

The most significant editorial comment (not a blocking issue) is in barbagroup/AmgXWrapper#17 which proposes a much more useful way for the software to be organized and distributed. Distributing as a plugin providing a PETSc PC would immediately allow AmgX to be used by any software that solves systems using PETSc.
