### 2017-08-08 #280 @danielskatz
@jedbrown - Can you check off the items that you think are done in the list above?

And make it clear what else you think is needed to check off the remaining items, so that the authors know what they need to do vs what they could do later.
