### 2017-08-09 barbagroup/AmgXWrapper#20 @jedbrown
Add CONTRIBUTING.md
... or otherwise document your contribution guidelines.
