### 2017-08-09 #280 @jedbrown
@danielskatz Done. Some comments:

* Performance: I don't have an equivalent system and can't verify specific performance claims, but I do see respectable performance from AmgX Classical.
* Installation: The instructions are adequate, but there is an implicit dependency on an antique version of GCC (via CUDA-6.5) and there is no package manager solution (though this is a sufficiently mixed stack that no single method would be platform independent). It would help to describe the variables `CUDA_DIR` and `AMGX_DIR` since (to my knowledge) these are not standard names so it can be unclear whether this should be a normal prefix (libraries in $AMGX_DIR/lib) or more complete path.
* Example: The example is very much a toy problem, but is relatively easy to understand. There is not currently an example of using the compiled library versus compiling the sources as part of the executable.
* Tests: There are no automated tests, but there is a manual process. See also barbagroup/AmgXWrapper#19.
* Community: There is no CONTRIBUTING.md or equivalent (see barbagroup/AmgXWrapper#20) but the project appears to follow de facto conventions for projects on GitHub.
