### 2017-08-09 #280 @danielskatz
Thanks @jedbrown  - I think you've done a nice job of saying where this meets a minimal set of requirements, where improvements could be made (but are not mandatory for JOSS acceptance), and where one change is needed for JOSS acceptance.
