### 2017-08-09 #280 @danielskatz
@piyueh - As you can see from the discussion, there is one change you need to make to allow JOSS to accept this - as discussed in https://github.com/barbagroup/AmgXWrapper/issues/20
