### 2017-08-09 barbagroup/AmgXWrapper#17 @piyueh
Agree. This is actually what we planned to do at the beginning: developing the interface directly in PETSc. In this way, we don't even have to change any code in PETSc applications.

But at that time, we were not familiar with advance PETSc usage, so while we needed something that we could quickly use in our application code, we decided to write this wrapper.

We'll definitely look into this. Do you have any suggestion about which example code we should look first?
