### 2017-08-09 2d2727f
resolve #13: applyNeumannBC and add problem specifiction in main.cpp
We rename the function `applyNeumannBC` to `fixSingularMat` to reflect
the fact that what we are doing in this function is pinning a point
with Dirichlet BC to resolve singular mat due to all-Neumann BC. All
Neumann BC is already incoporated into matrix A when we create it with
function `generateA`.
