+++ example/poisson/src/fixSingularMat.hpp
@@ -1,6 +1,6 @@
 /**
- * \file applyNeumannBC.hpp
- * \brief prototypes of applyNeumannBC.
+ * \file fixSingularMat.hpp
+ * \brief prototypes of fixSingularMat.
  * \author Pi-Yueh Chuang (pychuang@gwu.edu)
  * \date 2017-06-01
  */
@@ -15,7 +15,7 @@
 
 
 /**
- * \brief modify matrix A and RHS vector for all-Neumann BCs
+ * \brief modify matrix A and RHS vector for all-Neumann BCs to avoid sungular matrix.
  *
  * All-Neumann BC causes issues like indefinite matrix or infinite number of 
  * solutions. We hence assume the exact solution at the node i = j = 0 is known.
@@ -24,10 +24,12 @@
  * A[0, *] = A[*, 0] = 0; also all corresponding entries in RHS vector is
  * modified respectively.
  *
+ * In other words, we apply a Dirichlet BC at one point.
+ *
  * \param A [in, out] matrix A generated from the function generateA(...)
  * \param RHS [in, out] the right-hand-side vector generated from generateRHS(...)
  * \param exact [in] the exact solution vector generated from generateExt(...)
  *
  * \return PetscErrorCode.
  */
-PetscErrorCode applyNeumannBC(Mat &A, Vec &RHS, const Vec &exact);
+PetscErrorCode fixSingularMat(Mat &A, Vec &RHS, const Vec &exact);
