### 2017-08-09 e9fc2a0
resolve #14: remove exit(1) and printf
