### 2017-08-09 barbagroup/AmgXWrapper#17 @jedbrown
I would suggest starting with `src/ksp/pc/impls/saviennacl/saviennacl.cxx` in PETSc 'master', which interfaces to a (GPU-capable) smoothed aggregation solver in ViennaCL. @karlrupp could comment further about CUDA-specific details.
