### 2017-08-09 2b3ce46
resolve issuse #20: CONTRIBUTING.md
