### 2017-08-09 barbagroup/AmgXWrapper#20 @piyueh
Add CONTRIBUTING.md in commit 2b3ce465d1b1a9cc1dfef1b0cc8ecbfccc12ee33
