### 2017-08-09 barbagroup/AmgXWrapper#19 @piyueh
Agree. We will create some automatic tests and make both examples use the library.
