### 2017-08-09 f9c5eed
remove runtime warning about re-initialization; use error
It makes more sense to return an error instead of a runtime warning, when
users' code trying to initialize an instance that has already been initialized.
Also, this commit tries to resolve that, when the globalCpuWorld used in the
previous initialization is different from the second initialization, it
is possible the message will not correctly be displayed.
