### 2017-08-09 #280 @piyueh
@danielskatz @jedbrown
I've added **CONTRIBUTING.md**.
