### 2017-08-22 #280 @danielskatz
@jedbrown - please check the contributing and if it's ok, please check off the item in the checklist above, and then, if everything else seems good, let me know that you recommend the paper be published.  Otherwise, let me know what else you think needs to be done.
