### 2017-08-22 #280 @labarba
Yay!! We have full checkmarks!

Thank you, @jedbrown —we received lots of good, deep, constructive criticism from your review.
