### 2017-08-22 #280 @danielskatz
Thanks for the good review @jedbrown

@arfon, this is ready to go to you now for the next steps towards acceptance and publishing.
