### 2017-08-27 #280 @arfon
@piyueh - could you please merge this change to the paper metadata: https://github.com/barbagroup/AmgXWrapper/pull/21

Also, at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
