### 2017-08-29 #280 @jedbrown
> https://doi.org is preferred, but the earlier syntax dx.doi.org remains fully supported

https://www.doi.org/doi_handbook/3_Resolution.html
