### 2017-08-29 #280 @danielskatz
Good point.  @arfon, we should change all of these that we generate automatically
