### 2017-08-29 #280 @arfon
> Good point. @arfon, we should change all of these that we generate automatically

Yep, I'd managed to forget about that. I've added it to the backlog here e
