### 2017-09-07 d0e8d40
update README with JOSS DOI and how to cite
