### 2017-09-10 #280 @labarba
I see this as accepted 12 days ago, but cannot find it published. Is it in some limbo? (currently updating my CV for promotion application!)
