### 2017-09-10 #280 @arfon
@labarba see https://github.com/openjournals/joss-reviews/issues/280#issuecomment-325818312
