### 2017-10-13 #280 @labarba
Great news! NVIDIA has now released AmgX with an open source license:
https://github.com/NVIDIA/AMGX
